+++ src/main/kotlin/com/github/ivw/ezmode/config/EzModeRcParser.kt
@@ -115,7 +115,6 @@ object EzModeRcParser {
     val keyword = scanner.next()
     return when (keyword) {
       "write" -> WriteAction(scanner.restOfLine())
-      "writevar" -> WriteVarAction(scanner.restOfLine())
       "mode" -> KeyAction.ChangeMode(scanner.restOfLine())
       "idea" -> IdeKeyAction(scanner.restOfLine())
       "native" -> {