+++ src/main/kotlin/com/github/ivw/ezmode/config/keyactions/WriteAction.kt
@@ -7,13 +7,14 @@ import com.intellij.openapi.command.*
 data class WriteAction(val text: String) : KeyAction() {
   override fun perform(e: EzModeKeyEvent, onComplete: OnComplete?) {
     WriteCommandAction.runWriteCommandAction(e.project) {
-      e.editor.caretModel.runForEachCaret { caret ->
+      e.editor.caretModel.runForEachCaretIndexed { caret, caretIndex ->
+        val resolvedText = text.resolveVars(e, caretIndex, caret)
         if (caret.hasSelection()) {
-          e.editor.document.replaceString(caret.selectionStart, caret.selectionEnd, text)
+          e.editor.document.replaceString(caret.selectionStart, caret.selectionEnd, resolvedText)
           caret.removeSelection()
         } else {
-          e.editor.document.insertString(caret.offset, text)
-          caret.moveCaretRelatively(text.length, 0, false, true)
+          e.editor.document.insertString(caret.offset, resolvedText)
+          caret.moveCaretRelatively(resolvedText.length, 0, false, true)
         }
       }
     }